+++ src/main/java/com/dateplanner/user/dto/UserSocialLoginRequestDto.java
@@ -1,4 +1,4 @@
-package com.dateplanner.admin.user.dto;
+package com.dateplanner.user.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AccessLevel;
@@ -9,15 +9,9 @@ import lombok.extern.slf4j.Slf4j;
 @Slf4j(topic = "DTO")
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class UserSocialJoinRequestDto {
+public class UserSocialLoginRequestDto {
 
     @Schema(description = "인가 코드")
     private String code;
-
-    @Schema(description = "인증 제공자")
-    private String provider;
-
-    @Schema(description = "사용자 닉네임")
-    private String nickname;
 
 }
