+++ src/main/java/com/dateplanner/user/entity/User.java
@@ -1,4 +1,4 @@
-package com.dateplanner.admin.user.entity;
+package com.dateplanner.user.entity;
 
 import com.dateplanner.constant.entity.BaseTimeEntity;
 import com.fasterxml.jackson.annotation.JsonProperty;
